Move tests and mark as deprecated
